requirements.txt
anthropic>=0.40.0
google-genai>=0.3.0
supabase>=2.0.0
notion-client>=2.2.0
tweepy>=4.14.0
python-telegram-bot>=20.0
yfinance>=0.2.40
fredapi>=0.5.2
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
pydantic>=2.7.0
pyyaml>=6.0.1
jinja2>=3.1.0
tenacity>=8.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
black>=24.0.0
structlog>=24.0.0
detect-secrets>=1.4.0
